moved read content from file to util
